dah bisa nampilih manp dan ambil lokasi user

+++ database/migrations/2025_05_26_121056_create_data_warungs_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->decimal('longitude', 10, 7);
             $table->decimal('rating'); 
             $table->integer('price'); // 1 = murah, 2 = sedang, 3 = mahal
-            $table->integer('accessibility'); // kemudahan akses jalan dari 1-10
+            $table->decimal('accessibility'); // kemudahan akses jalan dari 1-10
             $table->timestamps();
         });
     }
